Peso Objeto tipo Float

+++ src/main/java/com/magenta/lab8_magenta/model/daos/EnemigoDao.java
@@ -39,7 +39,7 @@ public class EnemigoDao extends BaseDao {
                 objeto.setIdObjeto(rs.getInt("o.idObjeto"));
                 objeto.setNombreObjeto(rs.getString("nombreObjeto"));
                 objeto.setEfecto(rs.getString("efecto"));
-                objeto.setPeso(rs.getDouble("peso"));
+                objeto.setPeso(rs.getFloat("peso"));
                 enemigo.setObjeto(objeto);
 
                 ClaseEnemigo claseEnemigo = new ClaseEnemigo();
@@ -92,7 +92,7 @@ public class EnemigoDao extends BaseDao {
                     objeto.setIdObjeto(rs.getInt("o.idObjeto"));
                     objeto.setNombreObjeto(rs.getString("nombreObjeto"));
                     objeto.setEfecto(rs.getString("efecto"));
-                    objeto.setPeso(rs.getDouble("peso"));
+                    objeto.setPeso(rs.getFloat("peso"));
                     enemigo.setObjeto(objeto);
 
                     ClaseEnemigo claseEnemigo = new ClaseEnemigo();
@@ -229,7 +229,7 @@ public class EnemigoDao extends BaseDao {
                     objeto.setIdObjeto(rs.getInt("o.idObjeto"));
                     objeto.setNombreObjeto(rs.getString("nombreObjeto"));
                     objeto.setEfecto(rs.getString("efecto"));
-                    objeto.setPeso(rs.getDouble("peso"));
+                    objeto.setPeso(rs.getFloat("peso"));
                     enemigo.setObjeto(objeto);
 
                     ClaseEnemigo claseEnemigo = new ClaseEnemigo();
